pyproject.toml
[tool.poetry]
name = "arguing-agents"
version = "0.1.0"
description = ""
authors = ["Adnantium <adnantium@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.14"
langchain-openai = "^0.1.1"
langchain-experimental = "^0.0.56"
faiss-cpu = "^1.8.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
